+++ FacilityHorror-NWAPI/FacilityHorror-NWAPI.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <TargetFramework>net48</TargetFramework>
 	  <LangVersion>9.0</LangVersion>
+	  <RootNamespace>LightsOut-NWAPI</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
@@ -24,13 +25,16 @@
     <Reference Include="Assembly-CSharp-firstpass, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
       <HintPath>C:\Games\Steam\steamapps\common\SCP Secret Laboratory Dedicated Server\SCPSL_Data\Managed\Assembly-CSharp-firstpass.dll</HintPath>
     </Reference>
+    <Reference Include="NWAPIPermissionSystem">
+      <HintPath>..\..\_Dependencies\NWapi\NWAPIPermissionSystem.dll</HintPath>
+    </Reference>
     <Reference Include="UnityEngine.CoreModule, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null">
       <HintPath>C:\Games\Steam\steamapps\common\SCP Secret Laboratory Dedicated Server\SCPSL_Data\Managed\UnityEngine.CoreModule.dll</HintPath>
     </Reference>
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Northwood.PluginAPI" Version="12.0.0-rc.6" />
+    <PackageReference Include="Northwood.PluginAPI" Version="12.0.0" />
   </ItemGroup>
 
 </Project>
